refactor: stockCount 타입 Long으로 통일

+++ hub-service/src/main/java/com/boxoffice/hubservice/stocktransfer/dto/response/TransferPlanResponseDto.java
@@ -6,6 +6,6 @@ import java.util.UUID;
 public record TransferPlanResponseDto(
         UUID fromHubId,
         String fromHubName,
-        Integer totalStock,
+        Long totalStock,
         List<SuggestedTransferResponseDto> suggestedTransfers
 ) { }
